02-20 커뮤니티 신고처리 완료

+++ src/main/java/com/green/gogiro/community/model/CommunityReportDto.java
@@ -4,6 +4,7 @@ import lombok.Data;
 
 @Data
 public class CommunityReportDto {
-    private int iboard;
-    private int ireport;
+    private Long iboard;
+    private Long ireport;
 }
+
